# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Action Culture API
  description: |
    API backend for the EventCulture / Action Culture platform — an Algerian
    cultural heritage and events management application.
  version: 1.0.0
  contact:
    name: EventCulture Team
servers:
  - url: http://localhost:3001/api
    description: Development
  - url: https://api.actionculture.dz/api
    description: Production

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token sent in the Authorization header.
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie set after login / token refresh.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    UploadResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            filename:
              type: string
            originalName:
              type: string
            url:
              type: string
            size:
              type: integer
            mimetype:
              type: string
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
paths:
...
  /upload/image:
    post:
      tags: [Upload]
      summary: Upload a single image (authenticated)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid file
  /upload/image/public:
    post:
      tags: [Upload]
      summary: Upload a public image (registration)
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
  /upload/document:
    post:
      tags: [Upload]
      summary: Upload a document (authenticated)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  type: string
                  format: binary
      responses:
        '200':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
  /upload/multiple:
    post:
      tags: [Upload]
      summary: Upload multiple images (max 10)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
      responses:
        '200':
          description: Images uploaded
  /upload/{id}:
    get:
      tags: [Upload]
      summary: Get media info
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Media metadata
    delete:
      tags: [Upload]
      summary: Delete a media file
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Media deleted

  # =========================================================================
  # Admin — Dashboard
  # =========================================================================
...